requirements.txt
# Dependências principais
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Supabase
supabase==2.3.4
postgrest-py==0.13.2
storage3==0.7.4
realtime==1.0.4

# Validação e serialização
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilitários
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1

# Data e tempo
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Desenvolvimento
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1